refactored form per new principle
