+++ ui/documents/semantic-ui/js/documents/forms/deposit/FormAppLayout.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { Container } from "semantic-ui-react";
+import { BaseFormLayout } from "@js/oarepo_ui";
+import { NRDocumentValidationSchema } from "@nr";
+
+export const FormAppLayout = () => {
+  const formikProps = {
+    validationSchema: NRDocumentValidationSchema,
+  };
+  return (
+    <Container fluid>
+      <BaseFormLayout formikProps={formikProps} />
+    </Container>
+  );
+};
+export default FormAppLayout;
