+++ ui/documents/semantic-ui/js/documents/forms/deposit/index.js
@@ -1,8 +1,14 @@
-import { createFormAppInit } from "@js/oarepo_ui";
-import { DepositForm } from "./DepositForm";
+import { createFormAppInit, parseFormAppConfig } from "@js/oarepo_ui/forms";
+import FormAppLayout from "./FormAppLayout";
+import { DepositFormControlPanel } from "./DepositFormControlPanel";
+import { DepositFormFields } from "./DepositFormFields";
+const { formConfig } = parseFormAppConfig();
+const { overridableIdPrefix } = formConfig;
 
 export const componentOverrides = {
-  "Default.Form.FormApp.layout": DepositForm,
+  [`${overridableIdPrefix}.FormApp.layout`]: FormAppLayout,
+  [`${overridableIdPrefix}.FormFields.container`]: DepositFormFields,
+  [`${overridableIdPrefix}.FormActions.container`]: DepositFormControlPanel,
 };
 
 createFormAppInit({ componentOverrides: componentOverrides });
